refactor: Move Dockerflow behavior into a module
