+++ src/web/mod.rs
@@ -1,4 +1,5 @@
 //! Web authentication, handlers, and middleware
+pub mod dockerflow;
 pub mod extractors;
 pub mod handlers;
 pub mod middleware;
@@ -6,10 +7,4 @@ pub mod middleware;
 mod test;
 mod user_agent;
 
-// Known DockerFlow commands for Ops callbacks
-pub const DOCKER_FLOW_ENDPOINTS: [&str; 4] = [
-    "/__heartbeat__",
-    "/__lbheartbeat__",
-    "/__version__",
-    "/__error__",
-];
+pub use dockerflow::DOCKER_FLOW_ENDPOINTS;
